+++ CHANGELOG.md
@@ -5,9 +5,12 @@
 - Parser
   - **Breaking Change**: `ParserBuilder.checkSymbols` is now private, use `options` in `ParserBuilder.build` instead.
   - **Breaking Change**: Remove `AdvancedBuilder.expand`, make `AdvancedBuilder` extend `ParserBuilder`, they both implement `IParserBuilder`.
+  - **Breaking Change**: `Traverser` must return a value.
   - Add `ParserBuilder.use` to apply custom decorators with chain call.
   - Add `IParserBuilder.priority/leftSA/rightSA` for simplified conflict resolution.
   - `AdvancedBuilder.resolveRS/resolveRR` support `+*?|()`.
+- Fix
+  - Return `null` from `Traverser` will be transformed to `undefined`.
 - Versioned documentation.
 
 ## v0.7.0